back_cloud_t0/requirements.txt
python-multipart==0.0.7
setuptools==68.2.2
wheel==0.41.2
fastapi[all]
sqlalchemy
uvicorn[standard]
python-jose[cryptography]
psycopg2